+++ styles/globals.css
@@ -2,114 +2,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-body, html {
-  margin: 0;
-}
-
-body {
-  font-family: 'IBM Plex Mono', monospace;
-  padding: 50px;
-  background-color: #000;
-  color: #fff;
-}
-
-.container {
-  display: flex;
-  flex-flow: column;
-  align-items: center;
-}
-
-.title {
-  display: flex;
-  flex-flow: column;
-  align-items: center;
-  text-align: center;
-  gap: 8px;
-  margin-bottom: 32px;
-}
-
-.subhead {
-  font-size: 12px;
-  text-transform: uppercase;
-}
-
-.header {
-  font-size: 48px;
-}
-
-.form-input {
-  text-transform: uppercase;
-}
-
-.form-label {
-  margin-bottom: 10px;
-  align-self: center;
-}
-
-.form {
-  display: flex;
-  flex-flow: column;
-  align-items: flex-start;
-  text-align: center;
-  min-width: 500px;
-}
-.icon {
-  align-self: baseline;
-}
-
-input {
-  display: block;
-  border: none;
-  border: 1px solid #fff;
-  background: none;
-  color: white;
-  font-size: 16px;
-  height: 32px;
-  padding: 14px;
-  min-width: 250px;
-  font-family: 'IBM Plex Mono', monospace;
-  width: -webkit-fill-available;
-  width: -moz-available;
-  width: stretch;
-  margin-bottom: 12px;
-}
-
-input:focus {
-  outline: none;
-}
-
-/* hide default input arrows */
-input[type=number]::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-}
-
-.send {
-  border: 1px solid white;
-  background: none;
-  padding: 14px;
-  color: #999999;
-  cursor: pointer;
-  font-family: 'IBM Plex Mono', monospace;
-  width: 100%;
-  font-size: 16px;
-}
-
-.send:hover {
-  color: white;
-}
-
-.wallet-connect {
-  flex-direction: column;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 30px;
-}
-
-@media only screen and (max-width: 900px) {
-  input {
-    width: 250px;
-  }
-}
-
